Move unsafe loads to a shared spec
Psych 4.0 defaults `load` behavior to `safe_load` so these unsafe
loads need to be guarded on version. Safe versions should probably
be added.
